[GE-36] refactor: package명 변경
