+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.7'
 }
 
-group = 'com.menual'
+group = 'com.ceos.menual'
 version = '0.0.1-SNAPSHOT'
 description = 'Menual'
