+++ src/main/java/com/ceos/menual/MenualApplication.java
@@ -1,0 +1,13 @@
+package com.ceos.menual;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MenualApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MenualApplication.class, args);
+    }
+
+}
