+++ src/main/java/com/ceos/menual/domain/user/repository/UserRepository.java
@@ -1,0 +1,16 @@
+package com.ceos.menual.domain.user.repository;
+
+import com.ceos.menual.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+
+    boolean existsByNickname(String nickname);
+
+    Optional<User> findByEmail(String email);
+
+}
