+++ src/main/java/com/ceos/menual/entity/enums/UserType.java
@@ -1,0 +1,6 @@
+package com.ceos.menual.entity.enums;
+
+public enum UserType {
+    GROOMER,
+    EXPERT
+}
