+++ src/main/java/com/ceos/menual/global/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.ceos.menual.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
