+++ src/main/java/com/ceos/menual/global/exception/GlobalErrorCode.java
@@ -1,0 +1,17 @@
+package com.ceos.menual.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum GlobalErrorCode implements ResultCode{
+	//global
+	SUCCESS(HttpStatus.OK, 0, "정상 처리 되었습니다.");
+
+	private final HttpStatus status;
+	private final int code;
+	private final String message;
+}
