+++ src/main/java/com/ceos/menual/global/exception/ResultCode.java
@@ -1,0 +1,9 @@
+package com.ceos.menual.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ResultCode {
+	HttpStatus getStatus();
+	int getCode();
+	String getMessage();
+}
