--- src/main/java/com/groomeasy/backend/MenualApplication.java
@@ -1,13 +1,0 @@
-package com.groomeasy.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MenualApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MenualApplication.class, args);
-    }
-
-}
