--- src/main/java/com/groomeasy/backend/domain/user/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.groomeasy.backend.domain.user.repository;
-
-import com.groomeasy.backend.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    boolean existsByEmail(String email);
-
-    boolean existsByNickname(String nickname);
-
-    Optional<User> findByEmail(String email);
-
-}
