--- src/main/java/com/groomeasy/backend/entity/enums/UserType.java
@@ -1,6 +1,0 @@
-package com.groomeasy.backend.entity.enums;
-
-public enum UserType {
-    GROOMER,
-    EXPERT
-}
