--- src/main/java/com/groomeasy/backend/global/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.groomeasy.backend.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
