--- src/main/java/com/groomeasy/backend/global/exception/GlobalException.java
@@ -1,10 +1,0 @@
-package com.groomeasy.backend.global.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class GlobalException extends RuntimeException {
-	private final ResultCode resultCode;
-}
