--- src/main/java/com/groomeasy/backend/global/exception/ResultCode.java
@@ -1,9 +1,0 @@
-package com.groomeasy.backend.global.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ResultCode {
-	HttpStatus getStatus();
-	int getCode();
-	String getMessage();
-}
