+++ src/test/java/com/ceos/menual/MenualApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ceos.menual;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MenualApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
