--- src/test/java/com/groomeasy/backend/BackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.groomeasy.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
